+++ src/test/java/org/cyclops/commoncapabilities/ingredient/TestIngredientMatcherEnergy.java
@@ -53,69 +53,69 @@ public class TestIngredientMatcherEnergy {
     @Test
     public void testInstance() {
         assertThat(M.isInstance("a"), is(false));
-        assertThat(M.isInstance(123), is(true));
+        assertThat(M.isInstance(123L), is(true));
     }
 
     @Test
     public void testMatches() {
-        assertThat(M.matches(123, 124, true), is(false));
-        assertThat(M.matches(123, 123, true), is(true));
-        assertThat(M.matches(123, 124, false), is(true));
-        assertThat(M.matches(123, 123, false), is(true));
+        assertThat(M.matches(123L, 124L, true), is(false));
+        assertThat(M.matches(123L, 123L, true), is(true));
+        assertThat(M.matches(123L, 124L, false), is(true));
+        assertThat(M.matches(123L, 123L, false), is(true));
     }
 
     @Test
     public void testMatchesExactly() {
-        assertThat(M.matchesExactly(123, 124), is(false));
-        assertThat(M.matchesExactly(123, 123), is(true));
+        assertThat(M.matchesExactly(123L, 124L), is(false));
+        assertThat(M.matchesExactly(123L, 123L), is(true));
     }
 
     @Test
     public void testEmpty() {
-        assertThat(M.getEmptyInstance(), is(0));
+        assertThat(M.getEmptyInstance(), is(0L));
     }
 
     @Test
     public void testIsEmpty() {
-        assertThat(M.isEmpty(123), is(false));
-        assertThat(M.isEmpty(0), is(true));
+        assertThat(M.isEmpty(123L), is(false));
+        assertThat(M.isEmpty(0L), is(true));
     }
 
     @Test
     public void testHash() {
-        assertThat(M.hash(123), is(123));
-        assertThat(M.hash(0), is(0));
+        assertThat(M.hash(123L), is(123));
+        assertThat(M.hash(0L), is(0));
     }
 
     @Test
     public void testCopy() {
-        assertThat(M.copy(123), is(123));
-        assertThat(M.copy(0), is(0));
+        assertThat(M.copy(123L), is(123L));
+        assertThat(M.copy(0L), is(0L));
     }
 
     @Test
     public void testCompare() {
-        assertThat(M.compare(123, 124), is(-1));
-        assertThat(M.compare(124, 124), is(0));
-        assertThat(M.compare(125, 124), is(1));
+        assertThat(M.compare(123L, 124L), is(-1));
+        assertThat(M.compare(124L, 124L), is(0));
+        assertThat(M.compare(125L, 124L), is(1));
     }
 
     @Test
     public void testGetQuantity() {
-        assertThat(M.getQuantity(123), is(123L));
-        assertThat(M.getQuantity(124), is(124L));
-        assertThat(M.getQuantity(125), is(125L));
+        assertThat(M.getQuantity(123L), is(123L));
+        assertThat(M.getQuantity(124L), is(124L));
+        assertThat(M.getQuantity(125L), is(125L));
     }
 
     @Test
     public void testSetQuantity() {
-        assertThat(M.withQuantity(123, 234L), is(234));
-        assertThat(M.withQuantity(124, 235L), is(235));
+        assertThat(M.withQuantity(123L, 234L), is(234L));
+        assertThat(M.withQuantity(124L, 235L), is(235L));
     }
 
     @Test
     public void testGetMaximumQuantity() {
-        assertThat(M.getMaximumQuantity(), is((long) Integer.MAX_VALUE));
+        assertThat(M.getMaximumQuantity(), is((long) Long.MAX_VALUE));
     }
 
     @Test
